+++ maroonxdr/maroonx/recipes/sq/recipes_DYNAMICWAVECAL.py
@@ -25,13 +25,12 @@ def makeDynamicWavecal(p):
     p.addDQ()  # just placeholder until MX is in caldb
     p.overscanCorrect()
     p.correctImageOrientation()
-    # # p.addVAR(read_noise=True,poisson_noise=True)
+    p.addVAR(read_noise=True,poisson_noise=True)
     # # get and save wavelength solution (either static reference or frame's unique sim cal solved)
     # first perform echelle extraction of fibers
-    # p.darkSubtraction()
-    # p.extractStripes()  # gets relevant flat and dark to cut out frame's spectra
-    # p.optimalExtraction()  # does 2D to 1D conversion of cut out spectra,
-    # decide if optimal extraction is needed in fiber 5 and if the primitive needs adjusting to handle that
+    p.darkSubtraction()
+    p.extractStripes()  # gets relevant flat and dark to cut out frame's spectra
+    p.optimalExtraction()  # TODO: box extraction on all 5 frames
     # TODO: perform dynamic wavecal calculations on the extracted fibers
     p.storeProcessedArc(suffix='_dynamic_wavecal')
     return