+++ maroonxdr/maroonx/recipes/sq/recipes_STATICWAVECAL.py
@@ -24,14 +24,13 @@ def makeStaticWavecal(p):
     p.addDQ()  # just placeholder until MX is in caldb
     p.overscanCorrect()
     p.correctImageOrientation()
-    # # p.addVAR(read_noise=True,poisson_noise=True)
-    # # get and save wavelength solution (either static reference or frame's unique sim cal solved)
+    p.addVAR(read_noise=True,poisson_noise=True)
+    # get and save wavelength solution (either static reference or frame's unique sim cal solved)
     # first perform echelle extraction of fibers
-    p.darkSubtraction()
     p.extractStripes()  # gets relevant flat and dark to cut out frame's spectra
-    p.optimalExtraction()  # does 2D to 1D conversion of cut out spectra,
-    # decide if optimal extraction is needed in fiber 5 and if the primitive needs adjusting to handle that
+    p.optimalExtraction()  # TODO: box extraction on all 5 fibers
     # TODO: second perform static wavecal calculations on the extracted fibers
+    # 
     p.storeProcessedArc(suffix='_static_wavecal')
     return